+++ CHANGELOG.md
@@ -23,7 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed `StrEnum.from_str` with source as key ([#99](https://github.com/Lightning-AI/utilities/pull/99))
+- Fixed `StrEnum.from_str` with source as key (
+    [#99](https://github.com/Lightning-AI/utilities/pull/99),
+    [#102](https://github.com/Lightning-AI/utilities/pull/102)
+)
 
 
 ## [0.6.0] - 2023-01-23